move lmnet/utils/json.py to lmnet/utils/predict_output.py (#570)
